refactor(react): 拆出 helpers 模块，主循环剥离通用辅助函数
把 ReactEngine 主循环中不依赖引擎自身状态的纯过程性辅助函数收敛到
react/helpers.rs，使 engine.rs 聚焦于 ReAct 状态机本身。

迁移到 helpers.rs：
- looks_like_final_answer / FINAL_ANSWER_MIN_CHARS：最终回答启发式判断（含单测）
- drain_pending_commands_async：非阻塞命令排空（消息注入/取消/上下文压缩）
- maybe_inject_browser_update：节流浏览器页面观察与快照注入
- check_cancel：非阻塞取消/关闭检查

效果：engine.rs 从 1782 行降至 1497 行。至此 react 模块拆分完成，
engine.rs 相较重构前（2957 行）已减半，职责清晰收敛为 ReAct 状态机 +
工具调度，总结/取消/多代理/通用辅助各自独立成模块。行为无变化，208 测试全通过。
